Rename DecorationPoint
