+++ packages/slate-hyperscript/src/decorator-point.js
